don't use tmp when running --local
changed prefix to output_dir (default output_dir: none)

bump version

+++ server/src/tempdir.rs
@@ -1,34 +1,52 @@
-use std::process::Command;
-use log::{trace, debug, log_enabled};
+use log::{debug, log_enabled, trace};
 use std::ops::Deref;
+use std::{
+    path::{Path, PathBuf},
+    process::Command,
+};
 
 pub struct TempDir {
-    path: String,
+    path: PathBuf,
+    cleanup: bool,
 }
 
 impl TempDir {
+    /// Create a new TempDir.
+    /// Will delete the directory on drop.
     pub fn new() -> Self {
-        let tmpdir = String::from_utf8_lossy(
+        let output_str = String::from_utf8_lossy(
             &Command::new("mktemp")
                 .arg("-d")
                 .output()
                 .expect("Temporary file creation failed.")
                 .stdout, // removes the \n which mktemp appends
-        )
-        .to_string()
-        .split('\n')
-        .next()
-        .unwrap()
-        .to_owned();
+        ).to_string();
+
+        let tmpdir = output_str.split('\n').next().unwrap();
+
+        let tmpdir = PathBuf::from(tmpdir);
 
         debug!("Tempdir: {:?}", tmpdir);
 
-        Self { path: tmpdir }
+        Self {
+            path: tmpdir,
+            cleanup: true,
+        }
+    }
+
+    /// Create a new TempDir from an existing path.
+    /// Will not delete the directory on drop.
+    pub fn from_dir(path: &Path) -> Self {
+        let path = path.to_path_buf();
+        Self {
+            path,
+            cleanup: false,
+        }
     }
 }
 
 impl Deref for TempDir {
-    type Target = String;
+    type Target = PathBuf;
 
     fn deref(&self) -> &Self::Target {
         &self.path
@@ -37,11 +55,15 @@ impl Deref for TempDir {
 
 impl Drop for TempDir {
     fn drop(&mut self) {
+        if !self.cleanup {
+            return;
+        }
+
         trace!("Deleting files.");
         if log_enabled!(log::Level::Debug) {
             return;
         }
-        
+
         std::fs::remove_dir_all(&self.path).expect("Error deleting tmpfile.");
     }
 }
